requirements.txt
greenlet==2.0.2
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pynput==1.7.6
pyobjc-core==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Quartz==9.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.5.post1
tabulate==0.9.0
typing_extensions==4.5.0
